# Open files and store in dict for faster access