Tap peformance suffers when opening files in append mode repeatedly. This happens for each individual `RECORD` log line and so is very frequent. I've mocked up a quick time comparison in iPython to demonstrate the amount of overhead this creates.

```python
Python 3.8.5 (default, Aug  2 2020, 15:09:07)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.14.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: def test_reopen(fpath, lines):
   ...:     for l in lines:
   ...:         with open(fpath, 'a') as ostream:
   ...:            ostream.write(l + '\n')
   ...:

In [2]: def test_open_once(fpath, lines):
   ...:     with open(fpath, 'w') as ostream:
   ...:         for l in lines:
   ...:             ostream.write(l + '\n')
   ...:

In [3]: lines = list(map(str, range(0, 100)))

In [5]: %timeit test_reopen('test.txt', lines)
783 µs ± 9.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [4]: %timeit test_open_once('test.txt', lines)
92.7 µs ± 3.7 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

This yields an ~8.5x speed increase when the file contents can be written without re-opening in append mode.
 
The only downside to this approach is that the files must be closed before the process exits, as they are no longer opened and closed by the `with` statement.

## Changes

* Make a dict that holds `stream key -> opened output file` that is accessed in order to write lines to file
* Make helper method to ensure that all files are closed before exiting